fix issue: text offset

+++ src/text.c
@@ -99,7 +99,7 @@ void update_text(UIText * text) {
                 linear_coordinate = string_value[i] - 32;
             }
             offset_buffer[i].z = (GLfloat) (linear_coordinate % (int) DSTUDIO_CHAR_SIZE_DIVISOR) * (1.0 / DSTUDIO_CHAR_SIZE_DIVISOR);
-            offset_buffer[i].w = (GLfloat) (linear_coordinate / DSTUDIO_CHAR_SIZE_DIVISOR) * (1.0 / DSTUDIO_CHAR_SIZE_DIVISOR);
+            offset_buffer[i].w = (linear_coordinate / (int) DSTUDIO_CHAR_SIZE_DIVISOR) * (1.0 / DSTUDIO_CHAR_SIZE_DIVISOR);
         }
         text->actual_string_size = strlen(string_value);
         glBindBuffer(GL_ARRAY_BUFFER, text->instance_offsets);
